Updates to lightning-invoice man page
It all started when I saw the words "generating using".  So I decided to fix up
that section, and make sure I understood it.  Then I saw a few other nits
regarding commas and semi-colons so I fixed those.

This went on and on.

Found another error in man page reference:

* changed "list-invoice" to "list-invoices" since that is the correct name of the
  man page as it stands now.  But should the man page be list-invoice instead?

Other wordsmithing:

* changed "accepting payments" to "requesting payments" in name.
* changed "expectation of payment" to "payment request" in description.  This is how
  BOLT11 describes it.
* moved the fact that payment was in milli-satoshis from the main description to the
  'msatoshi' description.
* Added ERRORS section label
* underligned `lightningd` to be consistent with the lightning-cli man page.